Move the BasePipeline parts into its own module
