+++ scanpipe/pipelines/__init__.py
@@ -26,17 +26,10 @@ import traceback
 from contextlib import contextmanager
 from functools import wraps
 from pathlib import Path
-from pydoc import getdoc
-from pydoc import splitdoc
-from timeit import default_timer as timer
 
-from django.utils import timezone
-
-import bleach
-from markdown_it import MarkdownIt
 from pyinstrument import Profiler
 
-from scanpipe import humanize_time
+from pipeline import BasePipeline
 
 logger = logging.getLogger(__name__)
 
@@ -55,35 +48,72 @@ class InputFilesError(Exception):
         return message
 
 
-def group(*groups):
-    """Mark a function as part of a particular group."""
+class CommonStepsMixin:
+    """Common steps available on all project pipelines."""
 
-    def decorator(obj):
-        if hasattr(obj, "groups"):
-            obj.groups = obj.groups.union(groups)
-        else:
-            setattr(obj, "groups", set(groups))
-        return obj
+    def flag_empty_files(self):
+        """Flag empty files."""
+        from scanpipe.pipes import flag
 
-    return decorator
+        flag.flag_empty_files(self.project)
+
+    def flag_ignored_resources(self):
+        """Flag ignored resources based on Project ``ignored_patterns`` setting."""
+        from scanpipe.pipes import flag
+
+        if ignored_patterns := self.env.get("ignored_patterns"):
+            flag.flag_ignored_patterns(self.project, patterns=ignored_patterns)
+
+    def extract_archive(self, location, target):
+        """Extract archive at `location` to `target`. Save errors as messages."""
+        from scanpipe.pipes import scancode
+
+        extract_errors = scancode.extract_archive(location, target)
+
+        for resource_location, errors in extract_errors.items():
+            resource_path = Path(resource_location)
+
+            if resource_path.is_relative_to(self.project.codebase_path):
+                resource_path = resource_path.relative_to(self.project.codebase_path)
+                details = {"resource_path": str(resource_path)}
+            elif resource_path.is_relative_to(self.project.input_path):
+                resource_path = resource_path.relative_to(self.project.input_path)
+                details = {"path": f"input/{str(resource_path)}"}
+            else:
+                details = {"filename": str(resource_path.name)}
+
+            self.project.add_error(
+                description="\n".join(errors),
+                model="extract_archive",
+                details=details,
+            )
+
+    def extract_archives(self, location=None):
+        """Extract archives located in the codebase/ directory with extractcode."""
+        from scanpipe.pipes import scancode
+
+        if not location:
+            location = self.project.codebase_path
+
+        extract_errors = scancode.extract_archives(location=location, recurse=True)
+
+        for resource_path, errors in extract_errors.items():
+            self.project.add_error(
+                description="\n".join(errors),
+                model="extract_archives",
+                details={"resource_path": resource_path},
+            )
+
+        # Reload the project env post-extraction as the scancode-config.yml file
+        # may be located in one of the extracted archives.
+        self.env = self.project.get_env()
 
 
-def convert_markdown_to_html(markdown_text):
-    """Convert Markdown text to sanitized HTML."""
-    # Using the "js-default" for safety.
-    html_content = MarkdownIt("js-default").renderInline(markdown_text)
-    # Sanitize HTML using bleach.
-    sanitized_html = bleach.clean(html_content)
-    return sanitized_html
-
-
-class BasePipeline:
-    """Base class for all pipelines."""
+class Pipeline(CommonStepsMixin, BasePipeline):
+    """Main class for all project related pipelines including common steps methods."""
 
     # Flag specifying whether to download missing inputs as an initial step.
     download_inputs = True
-    # Flag indicating if the Pipeline is an add-on, meaning it cannot be run first.
-    is_addon = False
     # Optional URL that targets a view of the results relative to this Pipeline.
     # This URL may contain dictionary-style string formatting, which will be
     # interpolated against the project's field attributes.
@@ -92,150 +122,20 @@ class BasePipeline:
     results_url = ""
 
     def __init__(self, run):
-        """Load the Run and Project instances."""
-        self.run = run
-        self.project = run.project
-        self.pipeline_name = run.pipeline_name
+        """Load the Project instance."""
+        super().__init__(run)
+        self.project = self.run.project
         self.env = self.project.get_env()
 
-    @classmethod
-    def steps(cls):
-        raise NotImplementedError
-
-    @classmethod
-    def get_steps(cls, groups=None):
-        """
-        Return the list of steps defined in the ``steps`` class method.
-
-        If the optional ``groups`` parameter is provided, only include steps labeled
-        with groups that intersect with the provided list. If a step has no groups or
-        if ``groups`` is not specified, include the step in the result.
-        """
-        if not callable(cls.steps):
-            raise TypeError("Use a ``steps(cls)`` classmethod to declare the steps.")
-
-        steps = cls.steps()
-
-        if groups is not None:
-            steps = tuple(
-                step
-                for step in steps
-                if not getattr(step, "groups", [])
-                or set(getattr(step, "groups")).intersection(groups)
-            )
-
-        return steps
-
-    @classmethod
-    def get_doc(cls):
-        """Get the doc string of this pipeline."""
-        return getdoc(cls)
-
-    @classmethod
-    def get_graph(cls):
-        """Return a graph of steps."""
-        return [
-            {
-                "name": step.__name__,
-                "doc": getdoc(step),
-                "groups": getattr(step, "groups", []),
-            }
-            for step in cls.get_steps()
-        ]
-
-    @classmethod
-    def get_info(cls, as_html=False):
-        """Get a dictionary of combined information data about this pipeline."""
-        summary, description = splitdoc(cls.get_doc())
-        steps = cls.get_graph()
-
-        if as_html:
-            summary = convert_markdown_to_html(summary)
-            description = convert_markdown_to_html(description)
-            for step in steps:
-                step["doc"] = convert_markdown_to_html(step["doc"])
-
-        return {
-            "summary": summary,
-            "description": description,
-            "steps": steps,
-            "available_groups": cls.get_available_groups(),
-        }
-
-    @classmethod
-    def get_summary(cls):
-        """Get the doc string summary."""
-        return cls.get_info()["summary"]
-
-    @classmethod
-    def get_available_groups(cls):
-        return sorted(
-            set(
-                group_name
-                for step in cls.get_steps()
-                for group_name in getattr(step, "groups", [])
-            )
-        )
-
-    def log(self, message):
-        """Log the given `message` to the current module logger and Run instance."""
-        now_as_localtime = timezone.localtime(timezone.now())
-        timestamp = now_as_localtime.strftime("%Y-%m-%d %H:%M:%S.%f")[:-4]
-        message = f"{timestamp} {message}"
-        logger.info(message)
-        self.run.append_to_log(message)
-
-    @staticmethod
-    def output_from_exception(exception):
-        """Return a formatted error message including the traceback."""
-        output = f"{exception}\n\n"
-
-        if exception.__cause__ and str(exception.__cause__) != str(exception):
-            output += f"Cause: {exception.__cause__}\n\n"
-
-        traceback_formatted = "".join(traceback.format_tb(exception.__traceback__))
-        output += f"Traceback:\n{traceback_formatted}"
-
-        return output
-
-    def execute(self):
-        """Execute each steps in the order defined on this pipeline class."""
-        self.log(f"Pipeline [{self.pipeline_name}] starting")
-
-        steps = self.get_steps(groups=self.run.selected_groups)
-        selected_steps = self.run.selected_steps
-
-        if self.download_inputs:
-            steps = (self.__class__.download_missing_inputs,) + steps
-
-        steps_count = len(steps)
-        pipeline_start_time = timer()
-
-        for current_index, step in enumerate(steps, start=1):
-            step_name = step.__name__
-
-            if selected_steps and step_name not in selected_steps:
-                self.log(f"Step [{step_name}] skipped")
-                continue
-
-            self.run.set_current_step(f"{current_index}/{steps_count} {step_name}")
-            self.log(f"Step [{step_name}] starting")
-            step_start_time = timer()
-
-            try:
-                step(self)
-            except Exception as exception:
-                self.log("Pipeline failed")
-                return 1, self.output_from_exception(exception)
-
-            step_run_time = timer() - step_start_time
-            self.log(f"Step [{step_name}] completed in {humanize_time(step_run_time)}")
-
-        self.run.set_current_step("")  # Reset the `current_step` field on completion
-        pipeline_run_time = timer() - pipeline_start_time
-        self.log(f"Pipeline completed in {humanize_time(pipeline_run_time)}")
-
-        return 0, ""
+    # @classmethod
+    # def get_steps(cls, groups=None):
+    #     """Inject the ``download_inputs`` step if enabled."""
+    #     steps = super().get_steps(groups)
+    #
+    #     if cls.download_inputs:
+    #         steps = (cls.download_missing_inputs,) + steps
+    #
+    #     return steps
 
     def download_missing_inputs(self):
         """
@@ -295,67 +195,6 @@ class BasePipeline:
             yield
         except exceptions as error:
             self.add_error(exception=error, **kwargs)
-
-
-class Pipeline(BasePipeline):
-    """Main class for all pipelines including common step methods."""
-
-    def flag_empty_files(self):
-        """Flag empty files."""
-        from scanpipe.pipes import flag
-
-        flag.flag_empty_files(self.project)
-
-    def flag_ignored_resources(self):
-        """Flag ignored resources based on Project ``ignored_patterns`` setting."""
-        from scanpipe.pipes import flag
-
-        if ignored_patterns := self.env.get("ignored_patterns"):
-            flag.flag_ignored_patterns(self.project, patterns=ignored_patterns)
-
-    def extract_archive(self, location, target):
-        """Extract archive at `location` to `target`. Save errors as messages."""
-        from scanpipe.pipes import scancode
-
-        extract_errors = scancode.extract_archive(location, target)
-
-        for resource_location, errors in extract_errors.items():
-            resource_path = Path(resource_location)
-
-            if resource_path.is_relative_to(self.project.codebase_path):
-                resource_path = resource_path.relative_to(self.project.codebase_path)
-                details = {"resource_path": str(resource_path)}
-            elif resource_path.is_relative_to(self.project.input_path):
-                resource_path = resource_path.relative_to(self.project.input_path)
-                details = {"path": f"input/{str(resource_path)}"}
-            else:
-                details = {"filename": str(resource_path.name)}
-
-            self.project.add_error(
-                description="\n".join(errors),
-                model="extract_archive",
-                details=details,
-            )
-
-    def extract_archives(self, location=None):
-        """Extract archives located in the codebase/ directory with extractcode."""
-        from scanpipe.pipes import scancode
-
-        if not location:
-            location = self.project.codebase_path
-
-        extract_errors = scancode.extract_archives(location=location, recurse=True)
-
-        for resource_path, errors in extract_errors.items():
-            self.project.add_error(
-                description="\n".join(errors),
-                model="extract_archives",
-                details={"resource_path": resource_path},
-            )
-
-        # Reload the project env post-extraction as the scancode-config.yml file
-        # may be located in one of the extracted archives.
-        self.env = self.project.get_env()
 
 
 def is_pipeline(obj):
